QAレビュー指摘事項の全修正完了
## 修正内容

### 🚨 ブロッカー問題解決
- ✅ アーキテクチャ重複問題: `/lib/core/entities/` 削除、既存の `domain/entities` 統一
- ✅ Mock生成ファイル除外: `.gitignore` 追加、`**/*.mocks.dart` を管理対象外に
- ✅ 型整合性修正: Model クラスの `const` コンストラクタ → 通常コンストラクタ

### ⚠️ 重要問題対応
- ✅ Entity機能拡充: JSON変換、バリデーション、等価性比較を全Entity に追加
- ✅ バリデーション強化: 空文字チェック、範囲検証（緯度経度）、未来日時検証
- ✅ デフォルト値修正: `memo` フィールドのデフォルト空文字設定

### 🔧 コード品質向上
- ✅ JavaDoc完全化: 全Entityクラスに包括的ドキュメント追加
- ✅ テストパス統一: 全テストで `domain/entities` インポート使用
- ✅ アーキテクチャ準拠: Clean Architecture原則に完全準拠

## 品質確認
- ✅ **58テスト全て成功** (100%パス率達成)
- ✅ `flutter analyze` エラーゼロ
- ✅ `dart format` 準拠
- ✅ TDD Red→Green→Refactor完了
