+++ .gitignore
@@ -47,6 +47,9 @@ app.*.map.json
 # Coverage reports
 coverage/
 
+# Mock generated files
+**/*.mocks.dart
+
 # CI/CD artifacts
 *.apk
 *.ipa
